pipeline/requirements.txt
pandas
scikit-learn
category_encoders
joblib
omegaconf
hydra-core
